requirements.txt
numpy==1.21.5
opencv_python==4.7.0.68
Pillow==9.4.0
streamlit==1.18.1
torch==1.13.0
torchfile==0.1.0
torchvision==0.14.0
